[update] conform quoted string parameter to kernel
In kernel, if a parameter is given as:

'param=x y,z'

it will not parse this correctly as {"x y", "z"}.  However, it
supports quoting with " and without escape sequence.  Therefore, that
parameter should be sent as:

'param="x y,z"'

In user space then I adopted the same method solely for compatibility
with the kernel.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>