[update] change bool * to volatile sig_atomic_t *
This affects blocking lock functions where a bool * could have been used
to cancel the operation.  In compliance with the C standard, a
volatile sig_atomic_t * better suits the job as it can be set within a
signal handler.

sig_atomic_t in kernel space is typedefed as int in this commit as glibc
does.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>