[update] remove unshared stuff from shared memory
address, release and user_data were stored in the object momentarily to
avoid passing it to callback separately, which was not a very clean
thing.

What's more, with the fix in c8e8ead, the object size didn't
need to be updated on detach, so deregister doesn't get that anymore.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>