[build] upgrade build system to use autotools
This commit comes with some changes that are directly related to this
change of tools.

Note: the build system is still untested for RTAI.

With autotools, the user is much more free in deciding how to build and
install the software.  Therefore, a lot of the things that depend on the
user have been concealed, since they are not recoverable.  One such
example is urt-config --tool-*, since they need to find the installed
tool, which could be anywhere given the various user options that are
specified at install time rather than build time.

As good practice suggests keeping the configuration out of the header,
the include/urt_config.h file is removed.  Previously, only URT_SYS_*
was defined in it, which although possibly useful, was not very
necessary.  Instead of checking the definition from the header, one can
define what she wants from command line observing the output of
urt-config --sys.  include/urt_version.h is removed for a similar
reason.  If required, urt_version.h could be added again, or a method
for feature test could be introduced, but for now, this is still early
development and that's not quite necessary.

urt-config is in generally simplified.  In the end of the day, if it
could be completely removed, that would be the best, but currently, it
is still required.  Being installed in standard locations now means that
some of the include and link options are removed, but still the
subsystem-specific flags are retrieved through this script.

The build of documentation has been simplified because automake is very
picky about being portable (justifiably), which means GNU make functions
are not available, which makes for a very redundant Makefile.  Now it is
assumed that the builder doesn't manually delete _some_ of the generated
files and expect the build to make them.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>