[bugfix] [rtai] use rt_task_next_period
The urt_task_next_period function was broken for user space and it
needed to become exactly that implementation for posix.  However,  RTAI
should have really just had that function for user-space too.  So that's
what I did.  I just added that feature to RTAI in
github.com/ShabbyX/RTAI and now URT uses that.

Note: this is the first time URT really relies on the unofficial
features of RTAI in my repository, and since RTAI is being extremely
friendly with respect to patches, it's likely to remain this way
forever.  This means that URT won't be working with the official RTAI
repository.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>