[update] [rtai] disallow priority 0
This priority could be reserved for watchdog.  However, the actual
reason I did this is because a zero initialized urt_task_attr would have
this value set for priority, i.e. the highest possible priority which
could be dangerous.  Besides being dangerous, one would assume the
default priority should be used if no value is given (i.e. zero
initialized), which if 0 is allowed would not work that way.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>