[bugfix] mutually exclusive sysfs global sem operations
It turned out that if multiple threads of the same process try to `down`
the kernel semaphore through the sysfs interface at the same time, the
kernel blocks the whole process instead of just the thread.  I have yet
to identify the reason or the proper solution, but a workaround is to
have accesses to the sysfs file within the threads of the same process
to be mutually exclusive.

http://stackoverflow.com/q/28052643/912144

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>