[update] make urt_calibrate static
There was really no reason for this function to be visible to the user.
URT_GLUE took care of calling it before urt_init.  With this commit,
that function (now static in setup_common.c) is automatically called by
urt_init before anything else.

The end result is the same, except the user doesn't see urt_calibrate
anymore.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>