[update] remove (un)make_rt_context
This operation was not supported everywhere.  It was not a very
efficient operation to begin with and as it turned out also quite
unnecessary.  It is preferred for the programmer to take care of context
when calling RT or NONRT functions, than a library function itself to
try to switch to real-time context and back.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>