[update] make dist improvements
Some duplication is necessary for the sake of getting all of the files
to the distribution, regardless of which subsystem is being used.  That
is, where before there was `../sys/$(URT_SYS)`, there is now both
`../sys/posix` and `../sys/rtai`.  The rest of the changes are basically
putting scripts in EXTRA_DIST.

Oh, and urt.task_new in python doesn't use the variable `task` to avoid
confusion (by python) with the type urt.task.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>